+++ docs/Masa.Blazor.Docs/Examples/components/popup-service/Confirm.razor
@@ -71,13 +71,13 @@
                         throw new Exception("something wrong!");
                     }
 
-                    await PopupService.AlertAsync("Handle onOk successfully!", AlertTypes.Success);
+                    await PopupService.ShowSnackbarAsync("Handle onOk successfully!", AlertTypes.Success);
                 }
                 catch (Exception e)
                 {
                     args.Cancel();
 
-                    await PopupService.AlertAsync(e);
+                    await PopupService.ShowSnackbarAsync(e);
                 }
             });
     }
@@ -99,11 +99,11 @@
     {
         if (confirmed)
         {
-            await PopupService.AlertAsync("Ok was clicked!", AlertTypes.Info);
+            await PopupService.ShowSnackbarAsync("Ok was clicked!", AlertTypes.Info);
         }
         else
         {
-            await PopupService.AlertAsync("Cancel was clicked!", AlertTypes.Warning);
+            await PopupService.ShowSnackbarAsync("Cancel was clicked!", AlertTypes.Warning);
         }
     }
