+++ docs/Masa.Blazor.Docs/Examples/components/popup-service/Prompt.razor
@@ -48,7 +48,7 @@
                 {
                     if (args.Value == checkInput)
                     {
-                        await PopupService.AlertAsync($"Input('{args.Value}') is equal to '{checkInput}'", AlertTypes.Info);
+                        await PopupService.ShowSnackbarAsync($"Input('{args.Value}') is equal to '{checkInput}'", AlertTypes.Info);
                         
                         await Task.Delay(1000);
 
@@ -59,7 +59,7 @@
                     }
                     else
                     {
-                        await PopupService.AlertAsync($"Input('{args.Value}') is no equal to '{checkInput}'", AlertTypes.Error);
+                        await PopupService.ShowSnackbarAsync($"Input('{args.Value}') is no equal to '{checkInput}'", AlertTypes.Error);
                         args.Cancel();
                     }
                 }
@@ -67,7 +67,7 @@
                 {
                     args.Cancel();
 
-                    await PopupService.AlertAsync(e);
+                    await PopupService.ShowSnackbarAsync(e);
                 }
             });
     }
@@ -89,15 +89,15 @@
     {
         if (input is null)
         {
-            await PopupService.AlertAsync("Cancel was clicked", AlertTypes.Warning);
+            await PopupService.ShowSnackbarAsync("Cancel was clicked", AlertTypes.Warning);
         }
         else if (input == checkInput)
         {
-            await PopupService.AlertAsync($"Input('{input}') is equal to '{checkInput}'", AlertTypes.Info);
+            await PopupService.ShowSnackbarAsync($"Input('{input}') is equal to '{checkInput}'", AlertTypes.Info);
         }
         else
         {
-            await PopupService.AlertAsync($"Input('{input}') is no equal to '{checkInput}'", AlertTypes.Error);
+            await PopupService.ShowSnackbarAsync($"Input('{input}') is no equal to '{checkInput}'", AlertTypes.Error);
         }
     }
