+++ docs/Masa.Blazor.Docs/Examples/components/toasts/Duration.razor
@@ -1,12 +1,12 @@
-﻿<PToast @ref="_toast" Duration="duration">
-</PToast>
+﻿<PEnqueuedSnackbars @ref="_enqueuedSnackbars" Duration="duration">
+</PEnqueuedSnackbars>
 
-<MTextField Type="number" @bind-Value="duration" NumberProps="props=> { props.Min=0; props.Step=1000; }"></MTextField>
+<MTextField Type="number" @bind-Value="duration" NumberProps="props => { props.Min = 0; props.Step = 1000; }"></MTextField>
 
 <MButton OnClick="AddToast"> Add Toast </MButton>
 
 @code {
-    PToast _toast;
+    PEnqueuedSnackbars _enqueuedSnackbars;
 
     int? duration;
 
@@ -19,13 +19,14 @@
         var dark = random.NextInt64(0, 2);
         var config = new ToastConfig()
         {
-            Title = duration == 0? "Not auto close" : (duration == null? $"Auto close after 4000ms" : $"Auto close after {duration} ms"),
+            Title = duration == 0 ? "Not auto close" : (duration == null ? $"Auto close after 4000ms" : $"Auto close after {duration} ms"),
             Content = $"create time: {now}",
             Dark = dark == 0,
             Type = (AlertTypes)alertType,
             Duration = duration
         };
 
-        await _toast.AddToast(config);
+        await _enqueuedSnackbars.AddToast(config);
     }
+
 }
