+++ docs/Masa.Blazor.Docs/wwwroot/pages/components/popup-service/zh-CN.md
@@ -14,7 +14,7 @@ related:
 
 #### 警告
 
-<masa-example file="Examples.components.popup_service.Alert"></masa-example>
+<masa-example file="Examples.components.popup_service.Snackbar"></masa-example>
 
 #### 确认
