+++ src/Masa.Blazor/Popup/Components/EnqueuedSnackbars/EnqueuedSnackbars.razor
@@ -1,0 +1,9 @@
+﻿@using Masa.Blazor.Presets
+@namespace Masa.Blazor.Popup.Components
+
+<PEnqueuedSnackbars Position="_position"
+                   Duration="_duration"
+                   MaxCount="_maxCount"
+                   @ref="_toastRef">
+
+</PEnqueuedSnackbars>
