--- src/Masa.Blazor/Popup/Components/Toast/Toast.razor
@@ -1,6 +1,0 @@
-﻿@using Masa.Blazor.Presets
-@namespace Masa.Blazor.Popup.Components
-
-<PToast Position="_position" Duration="_duration" @ref="_toastRef" MaxCount="_maxCount">
-
-</PToast>
