+++ src/Masa.Blazor/Popup/PopupService.toast.cs
@@ -1,6 +1,5 @@
 ﻿using Masa.Blazor.Popup.Components;
 using Masa.Blazor.Presets;
-using OneOf;
 
 namespace Masa.Blazor
 {
