--- src/Masa.Blazor/Presets/Toast/PToast.razor
@@ -1,8 +1,0 @@
-﻿@namespace Masa.Blazor.Presets
-
-<div class="@CssProvider.GetClass()" style="@CssProvider.GetStyle()">
-    @foreach (var config in _configs)
-    {
-        <PToastItem Config="config" @key="config" OnClose="@HandleOnCloseAsync"></PToastItem>
-    }
-</div>
