+++ src/Masa.Blazor/wwwroot/css/masa-blazor.extend.css
@@ -484,43 +484,8 @@ input[type=number] {
         height: auto;
     }
 
-    .m-toast-container .m-snack__wrapper {
-        box-shadow: 0px 18px 40px rgba(112, 144, 176, 0.12);
-        border-radius: 14px
-    }
-
 .m-toast-item {
     align-items: flex-start;
-}
-
-    .m-toast-item .m-icon {
-        padding-top: 2px;
-    }
-
-.m-toast-detail {
-    font-size: 12px;
-    line-height: 18px;
-    color: #A3AED0;
-}
-
-.m-toast-title {
-    font-size: 16px;
-    line-height: 28px;
-    height: 28px;
-}
-
-.m-toast-text {
-    font-style: normal;
-    font-weight: 500;
-}
-
-.m-toast-container .m-snack__action {
-    padding-top: 9px;
-    align-self: flex-start;
-}
-
-.m-toast-container .m-snack__wrapper.theme--dark {
-    background-color: #323d6f
 }
 
 /* captcha */
@@ -738,4 +703,33 @@ input[type=number] {
     height: 100%;
     overflow: hidden;
     margin: 0;
+}
+
+/* alert */
+
+.m-alert__title {
+    align-items: center;
+    align-self: center;
+    display: flex;
+    font-size: 1.25rem;
+    font-weight: 500;
+    hyphens: auto;
+    letter-spacing: .0125em;
+    line-height: 1.5rem;
+    overflow-wrap: normal;
+    text-transform: none;
+    word-break: normal;
+    word-wrap: break-word;
+}
+
+/* snackbar > alert */
+
+.m-snack .m-alert {
+    padding: 0;
+    margin-bottom: 0;
+    background-color: transparent;
+}
+
+.m-snack .m-alert__icon {
+    margin-right: 12px;
 }
